Update User Command and Exception Handling
Added additional exception handling where exceptions might be thrown as well as implemented the update user command. User values will only update those that are supplied in the request sent to the server. This way, for example,  if client wants to only send an update of an x value change, that can be done.