#1 User Registration Add and Update
Users are now registered when the are added to the game. Their last update time is updated as well when they run an update command. Registration is set to inactive if activity is checked and they exceed timeout. Ticket #2 will handle automatic inactive user removal. get and list commands return user's activity status as well.